fix(bug): edit button overlapping text (#4276)
This pull request fixes the overlapping of the edit button on the
zaakgegevens fields and the location field. It also refactors the whole
of the zaakgegevens on the zaak-view html page. It now uses the [grid
layout](https://css-tricks.com/snippets/css/complete-guide-grid/#aa-introduction)
to show the fields in a better ordered way.



Solves PZ-8057