fix(zaak-create): only create class when value exists (#4306)
This pull request makes a small but important fix to the
`ZaakCreateComponent` to ensure that the `initiatorIdentificatie`
property is only set when a value is provided, preventing possible
errors when the value is missing.

- Updated the assignment of `initiatorIdentificatie` in the `createZaak`
method to set it to `null` if no value is provided, improving robustness
when handling optional form input.