fix: repeated initiator data calls (#4627)
This PR fixes the bug/function that when the KvK/Vestigingsnummer can't
be found/is a mismatch the code will retry this call multiple times,
even when it is a clear error. Also when the user switches tabs and
comes back to the zaak-view page the call was made. This is no longer
the case.

The logic is now that if the error status os above 500 or equal to 0 the
code may retry a maximum amount of 3 times to fetch the data.

`staleTime` has been introduced to disable the reload on each tab switch
but to stull reload the data after a certain period of time



Solves PZ-9071