feat: PABC use json import in docker compose (#4661)
This PR replaces the way the PABC database is filled. The old SQL files
are deleted and the new environment variable
`JSON_DATASET_PATH` is used to fill the database.

There are 2 new added files

- `pabc-mapping-data.json`: the actual JSON file with the data that
fills the database
- `pabc-mapping-data-schema.json`: the schema that the file need to
validated with

In the `docker-compose.yaml` the environment variable
`JSON_DATASET_PATH` has been added.

Per documentation and logic in `pabc-mapping-data-schema.json` when
`isAllEntityTypes` is true domainId can be null.
It appears that the logic for `isAllEntityTypes` doesn't work hence it
is removed when the domainId is null. Attempts to fill the database with
the `isAllEntityTypes: true` resulted in a exception thrown from PABC

Solves PZ-8037

---------

Co-authored-by: Edgar Vonk <edgar@info.nl>