fix: opening search after clicking doc detail page (#4920)
This PR fixes the bug that the search side nav doesn't get closed when
clicking on the detail page of a document.

Bug was caused by passing the zoekenSideNav property instead of calling
it as a function in the template; fixed by adding () so the MatSidenav
instance is passed correctly.

Solves PZ-9651 & PZ-9719

---------

Co-authored-by: Bas de Wit <52208549+bas-info-nl@users.noreply.github.com>